test: replaces test_multiplesubrprojects email test
