+++ tests/CMakeLists.txt
@@ -61,6 +61,7 @@ add_unit_test(/CDash/Model/BuildErrorFilter)
 add_unit_test(/CDash/Model/BuildFailure)
 add_unit_test(/CDash/Model/BuildRelationship)
 add_unit_test(/CDash/Model/Repository)
+add_unit_test(/CDash/MultipleSubprojectsEmail)
 add_unit_test(/CDash/Service/RepositoryService)
 add_unit_test(/CDash/ServiceContainer)
 add_unit_test(/CDash/TestUseCase)
